- Refactor model types from @/query to @/model
