+++ src/imp/api/filter/equals.ts
@@ -1,13 +1,9 @@
 import { OverpassEqualsFilter } from "@/imp/filter";
-import { OverpassExpression, OverpassFilter } from "@/model";
-import { OverpassFilterHelper } from "@/query";
+import { OverpassExpression, OverpassFilter, OverpassFilterHelper } from "@/model";
 import { OverpassFilterHelperBase } from "./base";
 
 export class OverpassEqualsFilterHelper extends OverpassFilterHelperBase {
-	constructor(
-		private readonly value: OverpassExpression<string>,
-		negated: boolean,
-	) {
+	constructor(private readonly value: OverpassExpression<string>, negated: boolean) {
 		super(negated);
 	}
