+++ src/model/index.ts
@@ -1,6 +1,8 @@
+export * from "./api";
 export * from "./compilable";
+export * from "./enum";
 export * from "./error";
 export * from "./expression";
-export * from "./filter";
-export * from "./statement";
+export * from "./parts";
+export * from "./query";
 export * from "./types";
